third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMorphologyElement-dom-radius-attr-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'radius' attribute of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.getAttribute('radius') is "6"
PASS morphologyElement.getAttribute('radius') is "4"
PASS successfullyParsed is true

TEST COMPLETE
